+++ frontend/src/hooks.js
@@ -1,8 +1,0 @@
-/** @type {import('@sveltejs/kit').Handle} */
-export async function handle({ event, resolve }) {
-    const response = await resolve(event, {
-        ssr: false
-    });
-    return response;
-}
-//# sourceMappingURL=hooks.js.map