+++ frontend/src/routes/components/ScriptBuilder.svelte
@@ -1,28 +1,28 @@
 <script lang="ts">
-	import { ScriptService, type Script } from '../../gen';
+	import { ScriptService, type Script } from '../../gen'
 
-	import { emptySchema, sendUserToast } from '../../utils';
-	import { onDestroy } from 'svelte';
-	import ScriptEditor from './ScriptEditor.svelte';
-	import { page } from '$app/stores';
-	import { goto } from '$app/navigation';
-	import Path from './Path.svelte';
-	import SvelteMarkdown from 'svelte-markdown';
-	import { workspaceStore } from '../../stores';
-	import ScriptSchema from './ScriptSchema.svelte';
-	import { inferArgs } from '../../infer';
-	import Required from './Required.svelte';
+	import { emptySchema, sendUserToast } from '../../utils'
+	import { onDestroy } from 'svelte'
+	import ScriptEditor from './ScriptEditor.svelte'
+	import { page } from '$app/stores'
+	import { goto } from '$app/navigation'
+	import Path from './Path.svelte'
+	import SvelteMarkdown from 'svelte-markdown'
+	import { workspaceStore } from '../../stores'
+	import ScriptSchema from './ScriptSchema.svelte'
+	import { inferArgs } from '../../infer'
+	import Required from './Required.svelte'
 
-	let editor: ScriptEditor;
-	let scriptSchema: ScriptSchema;
-	$: step = Number($page.url.searchParams.get('step')) || 1;
+	let editor: ScriptEditor
+	let scriptSchema: ScriptSchema
+	$: step = Number($page.url.searchParams.get('step')) || 1
 
-	export let script: Script;
-	export let initialPath: string = '';
+	export let script: Script
+	export let initialPath: string = ''
 
 	$: {
-		$page.url.searchParams.set('state', btoa(JSON.stringify(script)));
-		history.replaceState({}, '', $page.url);
+		$page.url.searchParams.set('state', btoa(JSON.stringify(script)))
+		history.replaceState({}, '', $page.url)
 	}
 
 	async function editScript(): Promise<void> {
@@ -38,47 +38,42 @@
 					schema: script.schema,
 					is_template: script.is_template
 				}
-			});
-			sendUserToast(`Success! New script version created with hash ${newHash}`);
-			goto(`/scripts/get/${newHash}`);
+			})
+			sendUserToast(`Success! New script version created with hash ${newHash}`)
+			goto(`/scripts/get/${newHash}`)
 		} catch (error) {
-			if (error.status === 400) {
-				sendUserToast(error.body, true);
-			} else {
-				sendUserToast(`Ooops.Something bad happened: ${error}`, true);
-				console.error(error);
-			}
+			sendUserToast(`Impossible to save the script: ${error.body}`, true)
 		}
 	}
 
 	export function setCode(script: Script) {
-		editor?.getEditor().setCode(script.content);
+		editor?.getEditor().setCode(script.content)
 
 		if (scriptSchema) {
 			if (script.schema) {
-				scriptSchema.setSchema(script.schema);
+				scriptSchema.setSchema(script.schema)
 			} else {
-				scriptSchema.setSchema(emptySchema());
+				scriptSchema.setSchema(emptySchema())
 			}
 		}
 	}
 
 	async function inferSchema() {
-		await inferArgs(script.content, script.schema);
+		await inferArgs(script.content, script.schema)
 	}
 
 	async function changeStep(step: number) {
 		if (step == 3) {
-			script.content = editor?.getEditor().getCode() ?? script.content;
-			await inferSchema();
-			script.schema = script.schema;
+			script.content = editor?.getEditor().getCode() ?? script.content
+			await inferSchema()
+			script.schema = script.schema
 		}
-		goto(`?step=${step}`);
+		goto(`?step=${step}`)
 	}
 
 	onDestroy(() => {
-		editor?.$destroy();
-	});
+		editor?.$destroy()
+	})
 </script>
 
 <div class="flex flex-col h-screen max-w-screen-lg xl:-ml-20 xl:pl-4 w-full -mt-4 pt-4 md:mx-10 ">
@@ -91,7 +86,7 @@
 						? 'default-button-disabled text-gray-700'
 						: 'default-button-secondary'} min-w-max ml-2"
 					on:click={() => {
-						changeStep(1);
+						changeStep(1)
 					}}>Step 1: Metadata</button
 				>
 				<button
@@ -99,7 +94,7 @@
 						? 'default-button-disabled text-gray-700'
 						: 'default-button-secondary'} min-w-max ml-2"
 					on:click={() => {
-						changeStep(2);
+						changeStep(2)
 					}}>Step 2: Code</button
 				>
 				<button
@@ -107,7 +102,7 @@
 						? 'default-button-disabled text-gray-700'
 						: 'default-button-secondary'} min-w-max ml-2"
 					on:click={() => {
-						changeStep(3);
+						changeStep(3)
 					}}>Step 3: UI customisation</button
 				>
 			</div>
@@ -118,15 +113,15 @@
 							(script.path == undefined || script.path == '' || script.path.split('/')[2] == '')}
 						class="default-button px-6 max-h-8"
 						on:click={() => {
-							changeStep(step + 1);
+							changeStep(step + 1)
 						}}>Next</button
 					>
 					{#if step == 2}
 						<button
 							class="default-button-secondary px-6 max-h-8 mr-2"
 							on:click={async () => {
-								await inferSchema();
-								editScript();
+								await inferSchema()
+								editScript()
 							}}>Save (commit)</button
 						>
 					{/if}