+++ frontend/src/routes/user/__layout-user@root.svelte
@@ -1,10 +1,9 @@
 <script lang="ts">
-	import { SvelteToast } from '@zerodevx/svelte-toast';
-	import { logout, sendUserToast } from '../../utils';
-	import { onMount } from 'svelte';
-	import { page } from '$app/stores';
-	import { superadmin, userStore, workspaceStore } from '../../stores';
-	import { UserService } from '../../gen';
+	import { SvelteToast } from '@zerodevx/svelte-toast'
+	import { logout, refreshSuperadmin, sendUserToast } from '../../utils'
+	import { onMount } from 'svelte'
+	import { page } from '$app/stores'
+	import { superadmin, userStore, workspaceStore } from '../../stores'
 
 	// Default options
 	const toastOptions = {
@@ -16,17 +15,17 @@
 		reversed: false, // insert new toast to bottom of stack
 		intro: { x: 256 }, // toast intro fly animation settings
 		theme: {} // css var overrides
-	};
+	}
 
 	$: {
 		if ($workspaceStore) {
-			localStorage.setItem('workspace', $workspaceStore);
+			localStorage.setItem('workspace', $workspaceStore)
 		}
 	}
 
 	onMount(() => {
 		window.onunhandledrejection = (e) => {
-			e.preventDefault();
+			e.preventDefault()
 			if (e.reason && e.reason.message) {
 				if (
 					['Model not found', 'Connection is disposed.', 'Connection got disposed.'].includes(
@@ -34,33 +33,25 @@
 					)
 				) {
 					// monaco editor promise cancelation
-					console.log('caught expected error');
+					console.log('caught expected error')
 				}
 				if (e.reason.status == '401') {
-					sendUserToast('Logged out after a request was unauthorized', true);
-					logout($page.url.pathname);
+					sendUserToast('Logged out after a request was unauthorized', true)
+					logout($page.url.pathname)
 				} else {
-					let message = `${e.reason?.message}: ${e.reason?.body ?? ''}`;
-					sendUserToast(message, true);
+					let message = `${e.reason?.message}: ${e.reason?.body ?? ''}`
+					sendUserToast(message, true)
 				}
 			} else {
-				console.log('unexpected error ignored', e);
+				console.log('unexpected error ignored', e)
 			}
-			e.preventDefault();
-			return false;
-		};
-		workspaceStore.set(undefined);
-		userStore.set(undefined);
-		if ($superadmin == undefined) {
-			UserService.globalWhoami().then((x) => {
-				if (x.super_admin) {
-					superadmin.set(x.email);
-				} else {
-					superadmin.set(false);
-				}
-			});
+			e.preventDefault()
+			return false
 		}
-	});
+		workspaceStore.set(undefined)
+		userStore.set(undefined)
+		refreshSuperadmin()
+	})
 </script>
 
 <div class="min-h-screen antialiased text-gray-900">