+++ frontend/src/routes/user/settings@user.svelte
@@ -1,33 +1,24 @@
 <script lang="ts">
-	import { userStore, workspaceStore } from '../stores';
+	import { usersWorkspaceStore } from '../../stores'
 
-	import { onMount } from 'svelte';
-	import type { TruncatedToken, NewToken } from '../gen';
-	import { UserService, SettingsService } from '../gen';
-	import { displayDate, sendUserToast, getToday } from '../utils';
-	import PageHeader from './components/PageHeader.svelte';
-	import CenteredPage from './components/CenteredPage.svelte';
-	import Icon from 'svelte-awesome';
-	import { faPlus } from '@fortawesome/free-solid-svg-icons';
-	import TableCustom from './components/TableCustom.svelte';
+	import type { TruncatedToken, NewToken } from '../../gen'
+	import { UserService, SettingsService } from '../../gen'
+	import { displayDate, sendUserToast, getToday } from '../../utils'
+	import PageHeader from './../components/PageHeader.svelte'
+	import Icon from 'svelte-awesome'
+	import { faPlus } from '@fortawesome/free-solid-svg-icons'
+	import TableCustom from '../components/TableCustom.svelte'
+	import CenteredModal from './CenteredModal.svelte'
 
-	let newPassword: string | undefined;
-	let passwordError: string | undefined;
-	let version: string | undefined;
-	let tokens: TruncatedToken[];
-	let newToken: string | undefined;
-	let newTokenLabel: string | undefined;
-	let newTokenExpiration: string | undefined;
-	let displayCreateToken = false;
-	let login_type = 'none';
-
-	function handleKeyUp(event: KeyboardEvent) {
-		const key = event.key || event.keyCode;
-		if (key === 13 || key === 'Enter') {
-			event.preventDefault();
-			setPassword();
-		}
-	}
+	let newPassword: string | undefined
+	let passwordError: string | undefined
+	let version: string | undefined
+	let tokens: TruncatedToken[]
+	let newToken: string | undefined
+	let newTokenLabel: string | undefined
+	let newTokenExpiration: string | undefined
+	let displayCreateToken = false
+	let login_type = 'none'
 
 	async function setPassword(): Promise<void> {
 		try {
@@ -36,73 +27,71 @@
 					requestBody: {
 						password: newPassword
 					}
-				});
-				sendUserToast('Your password was successfully updated');
+				})
+				sendUserToast('Your password was successfully updated')
 			} else {
-				sendUserToast('Specify a new password value to change your passord', true);
+				sendUserToast('Specify a new password value to change your passord', true)
 			}
 		} catch (error) {
-			sendUserToast(`Could not set this user's password: ${error}`, true);
+			sendUserToast(`Could not set this user's password: ${error}`, true)
 		}
 	}
 
 	async function loadVersion(): Promise<void> {
-		version = await SettingsService.backendVersion();
+		version = await SettingsService.backendVersion()
 	}
 	async function loadLoginType(): Promise<void> {
-		login_type = (await UserService.globalWhoami()).login_type;
+		login_type = (await UserService.globalWhoami()).login_type
 	}
 
 	async function createToken(): Promise<void> {
-		newToken = undefined;
-		let expirationISO: Date | undefined;
+		newToken = undefined
+		let expirationISO: Date | undefined
 		if (newTokenExpiration) {
-			expirationISO = new Date(newTokenExpiration);
+			expirationISO = new Date(newTokenExpiration)
 		}
 		try {
 			newToken = await UserService.createToken({
 				requestBody: { label: newTokenLabel, expiration: expirationISO?.toISOString() } as NewToken
-			});
-			listTokens();
-			displayCreateToken = false;
+			})
+			listTokens()
+			displayCreateToken = false
 		} catch (err) {
-			sendUserToast(`Could not create token: ${err}`, true);
+			sendUserToast(`Could not create token: ${err}`, true)
 		}
 	}
 
 	async function listTokens(): Promise<void> {
 		try {
-			tokens = await UserService.listTokens();
+			tokens = await UserService.listTokens()
 		} catch (err) {
-			sendUserToast(`Could not fetch tokens: ${err}`, true);
+			sendUserToast(`Could not fetch tokens: ${err}`, true)
 		}
 	}
 
 	async function deleteToken(tokenPrefix: string) {
 		try {
-			await UserService.deleteToken({ tokenPrefix });
-			sendUserToast('Succesfully deleted token');
-			listTokens();
+			await UserService.deleteToken({ tokenPrefix })
+			sendUserToast('Succesfully deleted token')
+			listTokens()
 		} catch (err) {
-			sendUserToast(`There was an error deleting this token: ${err}`, true);
+			sendUserToast(`There was an error deleting this token: ${err}`, true)
 		}
 	}
 
-	loadVersion();
-	loadLoginType();
-	$: {
-		if ($workspaceStore) {
-			listTokens();
-		}
-	}
+	loadVersion()
+	loadLoginType()
+	listTokens()
 </script>
 
-<CenteredPage>
-	<PageHeader title="User settings" />
+<CenteredModal title="User settings">
+	<div class="flex flex-row justify-between">
+		<a href="/user/workspaces">&leftarrow; Back to workspaces</a>
+	</div>
 	<div class="text-2xs text-gray-500 italic pb-6">
 		Running windmill version (backend) {version}
 	</div>
-	<h2 class="border-b">Change password</h2>
+	<h2 class="border-b">User info</h2>
 	<div class="">
 		{#if passwordError}
 			<div class="text-purple-500 text-2xs grow">{passwordError}</div>
@@ -110,10 +99,11 @@
 		<div class="flex flex-col gap-2 w-full ">
 			<div class="mt-4">
 				<label class="block w-60 mb-2 text-gray-500">
-					<input disabled value={$userStore?.email ?? ''} class="input mt-1" />
+					<div class="text-gray-700">email</div>
+					<input disabled value={$usersWorkspaceStore?.email} class="input mt-1" />
 				</label>
 				{#if login_type == 'password'}
-					<label class="block w-12/12">
+					<label class="block w-120">
 						<div class="text-gray-700">password</div>
 						<input
 							type="password"
@@ -131,6 +121,7 @@
 							text-sm
 							"
 						/>
+						<button on:click={() => setPassword()} class="mt-4 default-button">Set password</button>
 					</label>
 				{:else if login_type == 'github'}
 					<span>Authentified through Github OAuth2. Cannot set a password.</span>
@@ -144,10 +135,10 @@
 		<div class="text-right py-0 my-0 border-b">
 			<button
 				on:click={() => {
-					displayCreateToken = !displayCreateToken;
-					newToken = undefined;
-					newTokenExpiration = undefined;
-					newTokenLabel = undefined;
+					displayCreateToken = !displayCreateToken
+					newToken = undefined
+					newTokenExpiration = undefined
+					newTokenLabel = undefined
 				}}
 				class="default-button py-0 {displayCreateToken ? 'hidden' : ''}"
 				><Icon class="text-white mb-1" data={faPlus} scale={0.9} /> &nbsp; Create token</button
@@ -227,7 +218,7 @@
 								><button
 									class="text-red-500 text-xs underline"
 									on:click={() => {
-										deleteToken(token_prefix);
+										deleteToken(token_prefix)
 									}}>Delete</button
 								></td
 							>
@@ -243,7 +234,8 @@
 			</tbody>
 		</TableCustom>
 	</div>
-</CenteredPage>
 
-<style>
-</style>
+	<div class="flex flex-row justify-between pt-4">
+		<a href="/user/workspaces">&leftarrow; Back to workspaces</a>
+	</div>
+</CenteredModal>