+++ frontend/src/routes/user/workspaces@user.svelte
@@ -1,51 +1,53 @@
 <script lang="ts">
-	import { goto } from '$app/navigation';
-	import { page } from '$app/stores';
-	import { logout, logoutWithRedirect, sendUserToast } from '../../utils';
-	import { UserService, type WorkspaceInvite, WorkspaceService } from '../../gen';
-	import { superadmin, usersWorkspaceStore, workspaceStore } from '../../stores';
-	import CenteredModal from './CenteredModal.svelte';
-	import Switch from '../components/Switch.svelte';
+	import { goto } from '$app/navigation'
+	import { page } from '$app/stores'
+	import { logout, logoutWithRedirect, sendUserToast } from '../../utils'
+	import { UserService, type WorkspaceInvite, WorkspaceService } from '../../gen'
+	import { superadmin, usersWorkspaceStore, workspaceStore } from '../../stores'
+	import CenteredModal from './CenteredModal.svelte'
+	import Switch from '../components/Switch.svelte'
+	import { faCrown, faUser, faUserAlt, faUserCog } from '@fortawesome/free-solid-svg-icons'
+	import Icon from 'svelte-awesome'
 
-	let invites: WorkspaceInvite[] = [];
-	let list_all_as_super_admin: boolean = false;
-	let workspaces: { id: string; name: string; username: string }[] = [];
+	let invites: WorkspaceInvite[] = []
+	let list_all_as_super_admin: boolean = false
+	let workspaces: { id: string; name: string; username: string }[] = []
 
 	async function loadInvites() {
-		invites = await UserService.listWorkspaceInvites();
+		invites = await UserService.listWorkspaceInvites()
 	}
 
 	async function loadWorkspaces() {
 		if (!$usersWorkspaceStore) {
 			try {
-				usersWorkspaceStore.set(await WorkspaceService.listUserWorkspaces());
+				usersWorkspaceStore.set(await WorkspaceService.listUserWorkspaces())
 			} catch {}
 		}
 		if ($usersWorkspaceStore) {
 			if (!$workspaceStore) {
-				workspaceStore.set(localStorage.getItem('workspace')?.toString());
+				workspaceStore.set(localStorage.getItem('workspace')?.toString())
 			}
 		} else {
-			logoutWithRedirect($page.url.pathname);
+			logoutWithRedirect($page.url.pathname)
 		}
 	}
 
 	async function loadWorkspacesAsAdmin() {
-		workspaces = (await WorkspaceService.listWorkspacesAsSuperAdmin({})).map((x) => {
-			return { ...x, username: 'superadmin' };
-		});
+		workspaces = (await WorkspaceService.listWorkspacesAsSuperAdmin({ perPage: 1000 })).map((x) => {
+			return { ...x, username: 'superadmin' }
+		})
 	}
 
 	$: {
 		if (list_all_as_super_admin) {
-			loadWorkspacesAsAdmin();
+			loadWorkspacesAsAdmin()
 		} else {
-			workspaces = $usersWorkspaceStore?.workspaces ?? [];
+			workspaces = $usersWorkspaceStore?.workspaces ?? []
 		}
 	}
 
-	loadInvites();
-	loadWorkspaces();
+	loadInvites()
+	loadWorkspaces()
 </script>
 
 <CenteredModal title="Select a workspace" subtitle="Logged in as {$usersWorkspaceStore?.email}">
@@ -68,7 +70,8 @@
 			<button
 				class="
 					block
-					w-full
+					w-96
+					mx-auto
 					py-1
 					px-2
 					rounded-md
@@ -81,8 +84,8 @@
 					hover:ring-indigo-300
 					"
 				on:click={() => {
-					workspaceStore.set(workspace.id);
-					goto('/');
+					workspaceStore.set(workspace.id)
+					goto('/')
 				}}
 				><span class="font-mono">{workspace.id}</span> - {workspace.name} as
 				<span class="font-mono">{workspace.username}</span>
@@ -102,7 +105,8 @@
 		<div
 			class="
 					block
-					w-full
+					w-96
+					mx-auto
 					py-1
 					px-2
 					rounded-md
@@ -132,9 +136,9 @@
 					on:click={async () => {
 						await UserService.declineInvite({
 							requestBody: { workspace_id: invite.workspace_id }
-						});
-						sendUserToast(`Declined invite to ${invite.workspace_id}`);
-						loadInvites();
+						})
+						sendUserToast(`Declined invite to ${invite.workspace_id}`)
+						loadInvites()
 					}}
 				>
 					decline
@@ -142,11 +146,19 @@
 			</span>
 		</div>
 	{/each}
-	<div class="flex flex-row-reverse mt-10">
+	<div class="flex justify-between mt-10">
+		{#if $superadmin}
+			<a class="mr-10" href="/user/superadmin_settings">
+				<Icon data={faCrown} class="mr-1" scale={1} />Superadmin settings</a
+			>
+		{/if}
+		<a class="mr-10" href="/user/settings">
+			<Icon data={faUserCog} class="mr-1" scale={1} />User settings</a
+		>
 		<button
 			class="default-button-secondary"
 			on:click={async () => {
-				logout();
+				logout()
 			}}
 		>
 			logout